refactor(stats): 拆分StatsDashboard为概览卡片+控制栏+模型分布图
